fix(samv71): Fix integer overflow in PWMC duty cycle conversion
Problem:
- Pulse width not changing when using actuator_test commands
- ESC arms but motor doesn't respond to throttle changes

Root Cause:
- Integer overflow in microseconds-to-ticks conversion
- At MCK/64 (2,343,750 Hz): 2000µs * 2343750 = 4,687,500,000
- uint32_t max = 4,294,967,295 → OVERFLOW!
- Result wrapped to ~392 ticks instead of 4687 ticks

Solution:
- Use uint64_t for intermediate calculation in io_timer_set_ccr()
- Same fix for io_channel_get_ccr() reverse conversion

Math verification:
- 1000µs → 1000 * 2343750 / 1000000 = 2343 ticks (5% of 46875 period)
- 2000µs → 2000 * 2343750 / 1000000 = 4687 ticks (10% of 46875 period)

+++ platforms/nuttx/src/px4/microchip/samv7/io_pins/io_timer_pwmc.c
@@ -512,8 +512,10 @@ int io_timer_set_ccr(unsigned channel, uint16_t value)
 	uint32_t ch_base = get_channel_reg_base(channel);
 	uint8_t timer_idx = timer_io_channels[channel].timer_index;
 
-	/* Convert microseconds to timer ticks using current clock frequency */
-	uint32_t ticks = (uint32_t)value * g_timer_clock[timer_idx] / 1000000UL;
+	/* Convert microseconds to timer ticks using current clock frequency
+	 * Use uint64_t to prevent overflow: 2000 * 2343750 > uint32_t max!
+	 */
+	uint32_t ticks = (uint64_t)value * g_timer_clock[timer_idx] / 1000000ULL;
 
 	/* Clamp to period */
 	if (ticks > g_timer_period[timer_idx]) {
@@ -541,8 +543,10 @@ uint16_t io_channel_get_ccr(unsigned channel)
 	uint8_t timer_idx = timer_io_channels[channel].timer_index;
 	uint32_t ticks = pwm_ch_getreg(ch_base, PWM_CDTY_OFFSET);
 
-	/* Convert back to microseconds using current clock frequency */
-	return (uint16_t)(ticks * 1000000UL / g_timer_clock[timer_idx]);
+	/* Convert back to microseconds using current clock frequency
+	 * Use uint64_t to prevent overflow: 46875 * 1000000 > uint32_t max!
+	 */
+	return (uint16_t)((uint64_t)ticks * 1000000ULL / g_timer_clock[timer_idx]);
 }
 
 /**
